moved analytical LFP model functions etc. to file LFPy/test/common.py
